Extract shared methods from GeneratePomsMojo and its test class into an abstract superclass.
